Support .h1 in place of h1 inside .jumbotron
As in #7284, it can often be useful to style text like a heading in
places where using the heading tag would be either semantically
incorrect, less accessible, or otherwise invalid.

Support this in .jumbotron as in the rest of the document by allowing
elements with class .h1 to stand in for the h1 element.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>