Eliminate spurrious padding in .bs-docs-header
I agree with all the comments here.

Rather than dicker with container nesting at this late juncture, near sunset for BS 3.0x, let's just fix and move on.  It's docs only.

This PR addresses issue #15100 and eliminates the unnecessary horizontal scrollbar that appears for width for width >=768px and <= 782px.

Tested on three different iOS devices and FireFox/Chrome on OSX and it's fine now.

The only difference is the text in .bs-docs-header is shifted left by 15-pixels, lines are 30 px longer so wrap less in some circumstances in some viewports.