Change border class
The class ".border" (css/utilities/_borders.scss) is set the border-color to "$gray-200".
So I see that it would be nice to create a variable for the border color "$border-color". Because the default border should not always be "gray-200".
Another thing, there is a "$border-width" variable. Would not it be better to use this variable to set the default width of the border?